constants types refactoring

+++ src/include/Constants.h
@@ -7,8 +7,8 @@
 
 #include <cstdint>
 
-constexpr int kDefaultBacklog = 10;
-constexpr int kMaxFileNameSize = 4096;
+constexpr uint8_t kDefaultBacklog = 10;
+constexpr uint8_t kMaxFileNameSize = 4096;
 constexpr uint64_t kMaxFileSize = 10e12;
 
 #endif //CONSTANTS_H
